Include a Service for downloading the picture, even if the app is not running

+++ app/src/main/java/com/android/es/roversanz/series/data/DataPersistence.kt
@@ -6,7 +6,7 @@ interface DataPersistence {
 
     fun fetchSeries(): List<Serie>
 
-    fun fetchSeriesById(id: Long): Serie?
+    fun fetchSeriesById(id: Int): Serie?
 
     fun addSerie(serie: Serie)
