+++ app/src/main/java/com/android/es/roversanz/series/data/SerieRepository.kt
@@ -7,7 +7,7 @@ interface SerieRepository {
 
     fun getSeries(): Single<List<Serie>>
 
-    fun getSerie(id: Long): Single<Serie?>
+    fun getSerie(id: Int): Single<Serie?>
 
 }
 
@@ -20,7 +20,8 @@ class SerieRepositoryImpl(private val localDataPersistence: DataPersistence) : S
     override fun getSeries(): Single<List<Serie>> = Single.just(localDataPersistence.fetchSeries())
     //.delay(2, TimeUnit.SECONDS)
 
-    override fun getSerie(id: Long): Single<Serie?> = Single.just(localDataPersistence.fetchSeriesById(id))
+    override fun getSerie(id: Int): Single<Serie?> = Single.just(localDataPersistence
+                                                                       .fetchSeriesById(id))
     //
     // .delay(2, TimeUnit.SECONDS)
